Added ADD command
When the client wants to ADD one or more books to the list, there is a
three way handshake between the server and the client that happens.
After the user has been authenticated, the client can send as many as
books as he/she wants. After that the server appends these titles to the
bookList.txt and continues to wait for more commands.